pyproject.toml
[tool.poetry]
name = "components"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
eidolon-ai-sdk = "^0.1.116"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.24.0"
pytest-vcr = "^1.0.2"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
